tests for item types retrieval
