+++ src/main/resources/swagger.api/schemas/d2ir/centralItemTypesDTO.json
@@ -8,7 +8,7 @@
   "description": "D2IR central item types response",
   "type": "object",
   "properties": {
-    "ItemTypeList": {
+    "itemTypeList": {
       "description": "D2IR central item types",
       "type": "array",
       "items": {
