+++ src/test/java/org/folio/innreach/controller/base/BaseApiControllerTest.java
@@ -1,12 +1,29 @@
 package org.folio.innreach.controller.base;
 
+import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
+import static com.github.tomakehurst.wiremock.client.WireMock.equalTo;
+import static com.github.tomakehurst.wiremock.client.WireMock.stubFor;
+import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
 import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import static org.folio.innreach.fixture.TestUtil.readFile;
+
+import java.util.Collections;
+import java.util.Map;
 
 import javax.validation.Valid;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
+import com.github.tomakehurst.wiremock.client.MappingBuilder;
 import com.github.tomakehurst.wiremock.client.WireMock;
 import com.github.tomakehurst.wiremock.common.Slf4jNotifier;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
@@ -16,6 +33,8 @@ import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Profile;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.DynamicPropertyRegistry;
@@ -78,6 +97,46 @@ public class BaseApiControllerTest {
   @AfterEach
   void tearDown() {
     wm.resetAll();
+  }
+
+  protected void getAndExpect(String url, Template expectedResult) throws Exception {
+    mockMvc.perform(get(url))
+        .andExpect(status().isOk())
+        .andExpect(content()
+            .json(readTemplate(expectedResult)));
+  }
+
+  protected static void stubGet(String url, String responsePath) {
+    stubGet(url, Collections.emptyMap(), responsePath);
+  }
+
+  protected static void stubGet(String url, Map<String, String> requestHeaders, String responsePath) {
+    MappingBuilder getBuilder = WireMock.get(urlEqualTo(url));
+
+    requestHeaders.forEach((name, value) -> getBuilder.withHeader(name, equalTo(value)));
+
+    stubFor(getBuilder
+        .willReturn(aResponse()
+            .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
+            .withBodyFile(responsePath)));
+  }
+
+  private static String readTemplate(Template template) {
+    String path = "json/" + template.getFile();
+
+    return String.format(readFile(path), template.getParams());
+  }
+
+  @RequiredArgsConstructor(staticName = "of")
+  @AllArgsConstructor(access = AccessLevel.PRIVATE)
+  @Getter
+  protected static class Template {
+    final String file;
+    Object[] params;
+
+    public static Template of(String file, Object... params) {
+      return new Template(file, params);
+    }
   }
 
 }
