+++ src/test/resources/json/central-item-types/cs-item-types-empty-response.json
@@ -1,0 +1,4 @@
+{
+  "centralServerItemTypes": [],
+  "totalRecords": 0
+}
