+++ src/test/resources/wm/__files/error/d2ir-error-response.json
