+++ src/test/resources/wm/__files/item-types/d2ir-item-types-broken-response.json
@@ -1,0 +1,19 @@
+{
+  "status": "ok",
+  "reason": "success",
+  "errors": [],
+  "itemTypeList": [
+    {
+      "centralItemType": 200,
+      "description": "Book"
+    },
+    {
+      "centralItemType": 201,
+      "description": "Av"
+    },
+    {
+      "centralItemType": 202,
+      "description": Non-Circulation
+    }
+  ]
+}
