+++ src/test/resources/wm/__files/item-types/d2ir-item-types-response-01.json
@@ -1,0 +1,35 @@
+{
+  "status": "ok",
+  "reason": "success",
+  "errors": [],
+  "itemTypeList": [
+    {
+      "centralItemType": 200,
+      "description": "Book"
+    },
+    {
+      "centralItemType": 201,
+      "description": "Av"
+    },
+    {
+      "centralItemType": 202,
+      "description": "Non-Circulation"
+    },
+    {
+      "centralItemType": 210,
+      "description": "Requestable 4-week"
+    },
+    {
+      "centralItemType": 211,
+      "description": "Non-Requestable"
+    },
+    {
+      "centralItemType": 215,
+      "description": "Requestable 7-day"
+    },
+    {
+      "centralItemType": 217,
+      "description": "Requestable 7-day non-renew"
+    }
+  ]
+}
